Refactor MSDownload code to be unit tested later Added "esc" shortcut cancels AnswerDialog and AskDialog Fix centering on parent an screen for dialogs Made Myster's EventListener an interface Add DownloadInitiator to setup a NEW download Tweaked MSPartial file's MS download initiator

+++ src/main/java/com/myster/search/ui/SortableBit.java
@@ -11,7 +11,7 @@ public class SortableBit extends SortableLong {
     public String toString() {
         if (number < 0)
             return "??";
-        return "" + (((Long) getValue()).longValue() / 1000) + "Kbps";
+        return "" + (getValue().longValue() / 1000) + "Kbps";
     }
 
 }
